1. Push a old version since I last worked on project 2. Linux code untested 3. Some source code documention added 4. Couple descriptive variables added 5. Some PSP #ifdef code added 6. Emulator define changed to 0 to avoid some issues reported 7. Screen light toggle added, maybe not activated

+++ branches/cpp_rewrite/client/arm9/source/gui/rect.h
@@ -22,8 +22,9 @@ namespace D2K {
 				void SetW(uint8_t w);
 				uint8_t GetH();
 				void SetH(uint8_t h);
-				uint16_t GetX2();
-				uint16_t GetY2();
+				uint8_t GetX2();
+				uint8_t GetY2();
+				//@return true if (x), (y) Intersects, false if not
 				bool PointIntersect(uint8_t x, uint8_t y);
 				~Rect();
 		};
